requirements.txt
streamlit==1.45.1
openai==1.78.1
python-dotenv==1.1.0
supabase==1.0.3
st_supabase_connection==2.1.0